Begin separating ddtrace_wrap_fcall for PHP5/7
